requirements.txt
asgiref==3.8.0
Django==5.0.3
django-filter==24.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Markdown==3.6
PyJWT==2.8.0
sqlparse==0.4.4
tzdata==2024.1
